// src/entities/session/auth.type.ts
import { LearningTrack } from "../../shared/types/common.type";

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  fullName: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  userId: string;
  email: string;
  fullName: string;
  learningTrack: LearningTrack | null;
}